finish fixed tripPointList arch

+++ db/migrate/20120227052236_add_sort_id_to_trip_point.rb
@@ -1,5 +1,5 @@
 class AddSortIdToTripPoint < ActiveRecord::Migration
   def change
-    add_column :trip_points, :sort_id, :integer
+    add_column :trip_points, :sort_id, :float
   end
 end
